dabd-backend/requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
psycopg2-binary
python-dotenv
